# Conflicts:
#	Repository/Migrations/20220603094517_1.Designer.cs
#	Repository/Migrations/ProductDbContextModelSnapshot.cs

+++ DomainLayer/AdminSettings/Contact.cs
@@ -42,7 +42,7 @@ namespace DomainLayer.AdminSettings
         [RegularExpression(@"^\(?([0-9]{3})\)?[-. ]?([0-9]{3})[-. ]?([0-9]{4})$", ErrorMessage = "Entered phone format is not valid.")]
 
         [MaxLength(20)]
-        public long phoneNumber { get; set; }
+        public int phoneNumber { get; set; }
 
         [Required]
         [Column("email", TypeName = "nvarchar", Order = 8)]
